feat: 데미지 플로트 관련 스포너 생성 및 스크립트 정리

+++ Assets/02.Scripts/Ingame/Click/ClickInfo.cs
@@ -4,6 +4,6 @@ using UnityEngine;
 public struct ClickInfo
 {
     public EClickType ClickType;
-    public int Damage;
+    public double Damage;
     public Vector2 Position;
 }
